Ajustes de control al borrado de usuario admin y relación cursos-clases

+++ src/app/layouts/dashboard/pages/cursos/cursos.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ICurso } from './models'
+import { IClase } from '../clases/models';
 import { MatDialog } from '@angular/material/dialog';
 import { AbmCursosComponent } from './components/abm-cursos/abm-cursos.component';
 import { AuthService } from '../../../../core/services/auth.service';
 import { Subscription } from 'rxjs';
 import Swal from 'sweetalert2';
 import { CursosService } from './services/cursos.service';
+import { ClasesService } from '../clases/services/clases.service';
 
 @Component({
   selector: 'app-cursos',
@@ -22,10 +24,12 @@ export class CursosComponent implements OnInit{
 
   userData: Subscription =  new Subscription();
   cursos: ICurso[] = [];
+  clases: IClase[] = [];
   isAdmin: boolean = false;
 
   constructor(
     private cursosService: CursosService,
+    private clasesService: ClasesService,
     private matDialog: MatDialog,
     private authService: AuthService
   ) {}
@@ -77,24 +81,34 @@ export class CursosComponent implements OnInit{
       })
   };    
 
-  onDeleteCurso(id: number): void {
-    Swal.fire({
-      title: '¿Está seguro de eliminar el curso?',
-      icon: 'warning',
-      showCancelButton: true,
-    }).then((result) => {
-      if (result.isConfirmed) {
-        this.cursosService.deleteCurso(id).subscribe((data) => {
-          Swal.fire({
-            title: 'Curso eliminado',
-            icon: 'success',
-          });
-          this.cursos = data;
-        });
-      } else if (result.dismiss === Swal.DismissReason.cancel) {
+  onDeleteCurso(id: number, nombre: string): void {
+    this.clasesService.getClasesPorCurso(nombre).subscribe((clases) => {
+      if (clases.length > 0) {
         Swal.fire({
-          title: 'Petición Cancelada',
+          title: 'No se puede eliminar el curso',
+          text: 'Hay clases asociadas a este curso. Elimina las clases primero.',
           icon: 'error',
+        });
+      } else {
+        Swal.fire({
+          title: '¿Está seguro de eliminar el curso?',
+          icon: 'warning',
+          showCancelButton: true,
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.cursosService.deleteCurso(id).subscribe((data) => {
+              Swal.fire({
+                title: 'Curso eliminado',
+                icon: 'success',
+              });
+              this.cursos = data;
+            });
+          } else if (result.dismiss === Swal.DismissReason.cancel) {
+            Swal.fire({
+              title: 'Petición Cancelada',
+              icon: 'error',
+            });
+          }
         });
       }
     });
